Move non-preference preferences into new AppStatePersisted
This new class is for app state that needs to be maintained between runs of MT. It is not for user preferences despite
being built on `Preferences` as well.
