+++ src/main/java/net/rptools/maptool/client/AppSetup.java
@@ -157,7 +157,7 @@ public class AppSetup {
 
   public static void installLibrary(final String libraryName, final File root) {
     // Add as a resource root
-    AppPreferences.addAssetRoot(root);
+    AppStatePersisted.addAssetRoot(root);
     if (MapTool.getFrame() != null) {
       MapTool.getFrame().addAssetRoot(root);
